add idm agent (#125)
* add eval functions

* add idm agent

* rename the files

* add unit test for safety eval

* add unit test for efficiency metric

* minor

* minor

* add unit test for comfort evaluation

* minor

* idm unit test with minor fix

* add idm agent

* idm unit test with minor fix

* minor

* minor

* returns collided players

* minors + fix for python3.9

* improve test coverage

* use f-string

* bugfix

* add eval functions (#124)

Add common functions to evaluate the performance of ego vehicle in the simulation.
For safety, the functions to find the minimum distance, minimum time-to-collision(ttc) and maximum deceleration-rate-to-avoid-collision(drac) are implemented. The vehicle geometry is considered in the computation.
For efficiency, the function to find the first time that the ego vehicle entered its desired lane is implemented. It only checks whether the vehicle enters the goal lane(or its predecessor/successor), regardless of the progress on the lane.
For comfort, the max jerk and the rms of frequency-weighted acceleration(according to ISO2631) is computed.

This will be merged after tested locally for the pdm4ar exercise.
---------

Co-authored-by: alezana <azanardi@ethz.ch>

* revert change in eval

* revert change in eval

---------

Co-authored-by: alezana <azanardi@ethz.ch>